Teach japicmp about 'effectively final'
The JLS section on final methods says this:

  "Changing an instance method that is not declared final to be declared
  final may break compatibility with existing binaries that depend on the
  ability to override the method."

If the enclosing class is final, existing binaries have no mechanism to override
the method, so the change does not break compatibility.

This is relevant for Java to Kotlin migrations as the Kotlin compiler adds the
'final' modifier to all methods of final classes, whereas the Java compiler does
exactly what is in source code.

Closes: https://github.com/siom79/japicmp/issues/206